+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* Added `x-ydb-trace-id` header into grpc calls
+* Improved topic reader logs
+* Fixed `internal/xstring` package with deprecated warning in `go1.21` about `reflect.{String,Slice}Header`
+
 * Now read requests are sent after buffer free space reaches 30% of buffer size
 
 ## v3.51.3